+++ CHANGELOG.md
@@ -100,13 +100,13 @@
 - Added the `QUILL_DISABLE_FUNCTION_NAME` preprocessor flag and CMake option. This allows disabling `__FUNCTION__` in
   `LOG_*` macros when `%(caller_function)` is not used in `PatternFormatter`. This eliminates Clang-Tidy warnings when
   logging inside lambdas.
-- Added the `LOG_RUNTIME_METADATA(logger, log_level, file, line_number, function, message)` macro, which allows passing
+- Added the `LOG_RUNTIME_METADATA(logger, log_level, file, line_number, function, fmt, ...)` macro, which allows passing
   runtime metadata (such as file, line number, and function) along with a log message. This feature provides runtime
   flexibility, but it has a small overhead compared to the existing compile-time metadata macros. It is
   especially useful when forwarding logs received from another logging library to
   Quill. ([#696](https://github.com/odygrd/quill/issues/696))
   ```c++
-  LOG_RUNTIME_METADATA(logger, quill::LogLevel::Info, "main.cpp", 20, "foo", "Hello");
+  LOG_RUNTIME_METADATA(logger, quill::LogLevel::Info, "main.cpp", 20, "foo()", "Hello number {}", 8);
   ```
 - The `CsvWriter` could previously be used with `RotatingFileSink` via the constructor that accepted
   `std::shared_ptr<Sink>`, but rotated files did not include the CSV header. This has now been improved—when using the